+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning].
 ## Unreleased
 
 - fixed recipe viewer integration endpoints for Fabric
+- fixed unnecessary memory usage for debug handler
 
 ## [1.1.0] - 2024-09-27